[client] Remove range bounds in Bloom filter predicate
Lower and upper range bounds are required to allow merging
a range predicate with Bloom filter predicate on the same column.

Since the merging is done on the client side, the range bounds
are required in the wire protocol but we don't need them in
the public client APIs as client allows combining multiple predicates.

Change-Id: I60fa19e69ececa97e96eb6ed5f082406b0aab890