[util] Import "Or" function to BlockBloomFilter from Impala
Impala will be switching to using the Block Bloom filter from kudu-util.
"Or" function was missing and this change adds it.

Note that original implementation for OrEqualArrayAvx() in Impala is
targeted for AVX and not AVX2, however AVX2 is super-set
of AVX instructions and there is already provision in the Block
Bloom filter to separate out AVX2 v/s non-AVX2 (SSE) code. Hence
don't see need to add separate AVX specific file/implementation.

Change-Id: Ibe5f9311f73dcff883dd2cce18fd558e7d57d14f