[test] Start using GTEST_SKIP where applicable
Use the GTEST_SKIP() macro in SKIP_IF_SLOW_NOT_ALLOWED()
instead of early return so that test is marked
as SKIPPED instead of PASSED.

Replace instances where macro SKIP_IF_SLOW_NOT_ALLOWED()
is not used.

Use the GTEST_SKIP macro for a crypto test that is skipped
in FIPS mode.

On the dist-test UI, verified a skipped test is marked as
"successful" so no negative impact and same as earlier
but the test doesn't get marked as "skipped" or such.
Marking a test as skipped will require separate change to the
dist-test UI.

Sample output when tests are skipped:

[----------] Global test environment tear-down
[==========] 18 tests from 4 test suites ran. (27888 ms total)
[  PASSED  ] 14 tests.
[  SKIPPED ] 4 tests, listed below:
[  SKIPPED ] ParameterizedAddMasterTest.TestAddMasterCatchupFromWAL/0
[  SKIPPED ] ParameterizedAddMasterTest.TestAddMasterCatchupFromWAL/1
[  SKIPPED ] ParameterizedAddMasterTest.TestAddMasterSysCatalogCopy/0
[  SKIPPED ] ParameterizedAddMasterTest.TestAddMasterSysCatalogCopy/1

Change-Id: I4466bdde3965921d527131e76ad221643ff05705