[master] Script to automate adding a master
This script is a culmination of the changes that have been made so far
to add a master to a Kudu cluster.

This script must be run from the master being added and orchestrates
the necessary workflow with a bunch of verification.

This script will allow significantly simplifying the documentation for
adding a master to the cluster.

TODO:
- Test on a real cluster and MacOS
- Add more negative test cases or convert existing ones.

Change-Id: Ibc9dad73be625ff30070c522626bd79f1f1aa90c