[util] Add special handling for empty strings in FastHash
Impala uses special handling for empty strings while hashing
https://github.com/apache/impala/blob/master/be/src/runtime/raw-value.inline.h#L352

This leads to discrepancy between hash values for empty
strings between Impala and Kudu.
This change basically matches the behavior with Impala.
Since FastHash is not used for any persistent data structures
there are no backward compatibility issues with updating the
hashing implemenation.

Discovered bugs in the Java implementation when a negative
byte is promoted to a long and left shifted. Discovery
was because of the magic value for empty strings/buffer.
Thankfully the Java implementation of the FastHash
is not used currently.

Change-Id: Idb2d401ab692decd5bc03b48a96d710546c0039e