[docs][site] Add a note about Kudu version to doc pages.
Kudu documentation is specific to a Kudu version and we release
updated docs for every release. However we don't make it explicit that
the documentation applies to a specific version of Kudu.

This change basically adds a note about the Kudu version to most
docs by including a common file named top.adoc. This include
file can be used in future for other cases where a snippet needs
to be added in multiple files.

The docs that were left out didn't seem to be version specific and
hence didn't include the version note.

This'll apply the note to docs created henceforth for new releases and
existing docs won't get the update. In order to update the existing
docs for prior releases we'll have to update them manually.

Screenshots:
https://drive.google.com/file/d/1Tbbmbkj8KkCbtktSPsShUOHrox9PP39j/view
https://drive.google.com/file/d/11gyYJgR3jhHqKX6mXpvW9GGfnhb9b4_7/view

Upcoming change updates the landing page of releases linking the
versioned docs.

Change-Id: Ie391f32dd27b3464569d12b5b553c4ab84a95130